Comment out old BlogPage. Create new BlogPage container.
